zuwa-a-stock-analysis/requirements.txt
# 核心依赖
langchain>=0.3.0
langgraph>=0.2.0
openai>=1.0.0

# LLM API
openai

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# A股数据源
tushare>=1.4.0
akshare>=1.15.0

# 技术分析
ta-lib>=0.4.28

# Web UI
streamlit>=1.40.0
plotly>=5.18.0

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0

# 配置管理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.0

# 工具库
loguru>=0.7.0
pydantic>=2.5.0
python-dateutil>=2.8.0

# 飞书SDK
lark-oapi>=1.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
